Renamed probability to locate and updated throughout
